Put profiles in a single file.
