+++ tests/conftest.py
@@ -1,7 +1,6 @@
 import pytest
-import os
 
-from .utils import build_databricks_cluster_profile
+from tests.profiles import get_databricks_cluster_target
 
 
 pytest_plugins = ["dbt.tests.fixtures.project"]
@@ -23,51 +22,7 @@ def pytest_configure(config):
 @pytest.fixture(scope="session")
 def dbt_profile_target(request):
     profile_type = request.config.getoption("--profile")
-    if profile_type == "databricks_cluster":
-        target = databricks_cluster_target()
-    elif profile_type == "databricks_sql_endpoint":
-        target = databricks_sql_endpoint_target()
-    elif profile_type == "databricks_uc_cluster":
-        target = databricks_uc_cluster_target()
-    elif profile_type == "databricks_uc_sql_endpoint":
-        target = databricks_uc_sql_endpoint_target()
-    else:
-        raise ValueError(f"Invalid profile type '{profile_type}'")
-    return target
-
-
-def databricks_cluster_target():
-    return build_databricks_cluster_profile(
-        http_path=os.getenv(
-            "DBT_DATABRICKS_CLUSTER_HTTP_PATH", os.getenv("DBT_DATABRICKS_HTTP_PATH")
-        )
-    )
-
-
-def databricks_sql_endpoint_target():
-    return build_databricks_cluster_profile(
-        http_path=os.getenv(
-            "DBT_DATABRICKS_ENDPOINT_HTTP_PATH", os.getenv("DBT_DATABRICKS_HTTP_PATH")
-        )
-    )
-
-
-def databricks_uc_cluster_target():
-    return build_databricks_cluster_profile(
-        http_path=os.getenv(
-            "DBT_DATABRICKS_UC_CLUSTER_HTTP_PATH", os.getenv("DBT_DATABRICKS_HTTP_PATH")
-        ),
-        catalog=os.getenv("DBT_DATABRICKS_UC_INITIAL_CATALOG", "main"),
-    )
-
-
-def databricks_uc_sql_endpoint_target():
-    return build_databricks_cluster_profile(
-        http_path=os.getenv(
-            "DBT_DATABRICKS_UC_ENDPOINT_HTTP_PATH", os.getenv("DBT_DATABRICKS_HTTP_PATH")
-        ),
-        catalog=os.getenv("DBT_DATABRICKS_UC_INITIAL_CATALOG", "main"),
-    )
+    return get_databricks_cluster_target(profile_type)
 
 
 @pytest.fixture(autouse=True)
